+++ tests/testthat/test-utils.R
@@ -10,7 +10,7 @@ test_that("markdown chunks are created correctly",{
 
   png_obj <- png_path(path = system.file("extdata/test_image.png", package = "docorator"))
   docorator <- as_docorator(png_obj, display_name = "myfig", save_object = FALSE)
-  chunk_png <- create_chunk(x = docorator, transform = NULL) |> capture.output()
+  chunk_png <- create_chunks_all(x = docorator, transform = NULL) |> capture.output()
   # avoiding snapshot because of temp directory + mardown header
   expect_true(any(grepl('<div class=\"figure\">', chunk_png)))
 
@@ -23,13 +23,32 @@ test_that("markdown chunks are created correctly",{
                 tbl_scale = TRUE,
                 tbl_stub_pct = 0.5)
   docorator <- as_docorator(my_gt, display_name = "mytbl", save_object = FALSE)
-  chunk_gt <- create_chunk(x = docorator, transform = NULL) |> capture.output()
+  chunk_gt <- create_chunks_all(x = docorator, transform = NULL) |> capture.output()
   expect_snapshot(chunk_gt)
 
   my_gt_group <- gt::gt_group(my_gt,my_gt)
   docorator <- as_docorator(my_gt_group, display_name = "mytbl", save_object = FALSE)
   chunk_gt_group <- create_chunk(x = docorator, transform = NULL) |> capture.output()
   expect_snapshot(chunk_gt_group)
+
+
+  # list of ggplot2 - tempdir to stop creation of figure folder
+  withr::with_tempdir({
+
+  ggplot1 <- ggplot2::ggplot(data = mtcars, ggplot2::aes(y=cyl, x=mpg)) +
+    ggplot2::geom_point() +
+    ggplot2::labs(title = "title1", subtitle = "subtitle1", tag = "tag1", caption = "footnote1")
+  ggplot2 <- ggplot2::ggplot(data = mtcars, ggplot2::aes(x=cyl, y=mpg)) +
+    ggplot2::geom_point() +
+    ggplot2::labs(title = "title2", subtitle = "subtitle2", tag = "tag2", caption = "footnote2")
+
+  list_of_ggplots <- list(ggplot1, ggplot2)
+  docorator <- as_docorator(list_of_ggplots, display_name = "my_plot_list", save_object = FALSE)
+  chunk_list <- create_chunks_all(x = docorator, transform = NULL) |> capture.output()
+  expect_true(any(grepl('new_chunk1', chunk_list)))
+  expect_true(any(grepl('new_chunk2', chunk_list)))
+
+})
 
 })
 
@@ -67,4 +86,3 @@ test_that("package version messages are printed correctly",{
   })
 
 })
-